Check for dirs before rmdir
Fixes #158

+++ src/manifest.ts
@@ -571,9 +571,11 @@ export class Manifest {
     }
 
     async deleteRepoDir(): Promise<void> {
-        await fs.rmdir(this.repoDir, {
-            recursive: true,
-        })
+        if (await exists(this.repoDir)) {
+            await fs.rmdir(this.repoDir, {
+                recursive: true,
+            })
+        }
     }
 
     async overrideWorkspaceCommandConfig(
